Fix innerHTML security concern
Menu information, titles, and dates were changed using the innerHTML
attribute of each element. This was changed to textContent instead.
Tags were parsed using a DOMParser for elements that had HTML tags
built in. Support for Firefox was added as well; scrolling was
possible on the popup on Firefox and this was remedied by adding an
overflow: hidden line on the unorderered list tag in the styles.css
file.

Signed-off-by: Alex Le <alex.tle63@gmail.com>