Update 3.0.0
UI of pop is overgoing overhaul making experience smoother, it is
a work in progress. The height of the popup has been made static
and a new overlay menu to access tag preferences has been added.

The SwatDiningLibrary object has been fully implemented, and the
code was vastly cleaned up as a result. Updating the object
dynamically may pose an issue between different computers, but
the time set to render the object will be lengthened upon release.

The tag visibility switch will remain but the 'Edit Tags' menu
will be a new addition to accompany it. Users will be able to turn
on and off tags that are ir/relevant to their own personal use.
This metadata will be stored locally, but this is a feature that
is yet to be implemented.